fix: refactor cli-text-reader
