+++ cli-text-reader/src/tutorial.rs
@@ -1,0 +1,16 @@
+
+pub fn get_tutorial_text() -> Vec<String> {
+    vec![
+        "Welcome to the Text Reader!".to_string(),
+        "".to_string(),
+        "Basic Controls:".to_string(),
+        "  j or ↓    : Move down one line".to_string(),
+        "  k or ↑    : Move up one line".to_string(),
+        "  PageDown  : Move down one page".to_string(),
+        "  PageUp    : Move up one page".to_string(),
+        "  :z        : Toggle line highlighter".to_string(),
+        "  :q        : Quit".to_string(),
+        "".to_string(),
+        "Press any key to continue...".to_string(),
+    ]
+}
